create new integration testsuite for manifest manager with oci secret
