+++ tests/integration/controller/manifest/oci_reg_secret/suite_test.go
@@ -13,10 +13,11 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-package custom_resource_check_test
+package oci_reg_secret
 
 import (
 	"context"
+	"k8s.io/apimachinery/pkg/types"
 	"net/http/httptest"
 	"os"
 	"path/filepath"
@@ -50,7 +51,6 @@ import (
 	"github.com/kyma-project/lifecycle-manager/internal/manifest"
 	"github.com/kyma-project/lifecycle-manager/internal/manifest/img"
 	"github.com/kyma-project/lifecycle-manager/internal/manifest/manifestclient"
-	"github.com/kyma-project/lifecycle-manager/internal/manifest/statecheck"
 	"github.com/kyma-project/lifecycle-manager/internal/pkg/metrics"
 	kymarepository "github.com/kyma-project/lifecycle-manager/internal/repository/kyma"
 	"github.com/kyma-project/lifecycle-manager/internal/setup"
@@ -121,13 +121,6 @@ var _ = BeforeSuite(func() {
 	if !found {
 		metricsBindAddress = ":0"
 	}
-	cacheOpts := setup.SetupCacheOptions(false,
-		"istio-system",
-		ControlPlaneNamespace,
-		certmanagerv1.SchemeGroupVersion.String(),
-		logr)
-	syncPeriod := 2 * time.Second
-	cacheOpts.SyncPeriod = &syncPeriod
 
 	mgr, err = ctrl.NewManager(
 		cfg, ctrl.Options{
@@ -135,11 +128,13 @@ var _ = BeforeSuite(func() {
 				BindAddress: metricsBindAddress,
 			},
 			Scheme: k8sclientscheme.Scheme,
-			Cache:  cacheOpts,
+			Cache: setup.SetupCacheOptions(false,
+				"istio-system",
+				ControlPlaneNamespace,
+				certmanagerv1.SchemeGroupVersion.String(),
+				logr),
 		},
 	)
-
-	mgr.GetControllerOptions()
 	Expect(err).ToNot(HaveOccurred())
 
 	authUser, err := testEnv.AddUser(
@@ -151,26 +146,30 @@ var _ = BeforeSuite(func() {
 	Expect(err).NotTo(HaveOccurred())
 
 	kcpClient = mgr.GetClient()
-	keyChainLookup := keychainprovider.NewDefaultKeyChainProvider()
-	statefulChecker := statecheck.NewStatefulSetStateCheck()
-	deploymentChecker := statecheck.NewDeploymentStateCheck()
+	nonExistingSecretName := types.NamespacedName{Namespace: "kcp-system", Name: "non-existing-secret"}
+	keyChainLookup := keychainprovider.NewFromSecretKeyChainProvider(kcpClient, nonExistingSecretName)
 	extractor := img.NewPathExtractor()
 	testEventRec := event.NewRecorderWrapper(mgr.GetEventRecorderFor(shared.OperatorName))
 	manifestClient := manifestclient.NewManifestClient(testEventRec, kcpClient)
 	orphanDetectionClient := kymarepository.NewClient(kcpClient)
-	reconciler = declarativev2.NewFromManager(mgr, queue.RequeueIntervals{
-		Success: 1 * time.Second,
-		Busy:    1 * time.Second,
-		Error:   1 * time.Second,
-		Warning: 1 * time.Second,
-	}, metrics.NewManifestMetrics(metrics.NewSharedMetrics()), metrics.NewMandatoryModulesMetrics(),
-		manifestClient, orphanDetectionClient, spec.NewResolver(keyChainLookup, extractor),
+	reconciler = declarativev2.NewFromManager(mgr,
+		queue.RequeueIntervals{
+			Success: 1 * time.Second,
+			Busy:    1 * time.Second,
+			Error:   1 * time.Second,
+			Warning: 1 * time.Second,
+		},
+		metrics.NewManifestMetrics(metrics.NewSharedMetrics()),
+		metrics.NewMandatoryModulesMetrics(),
+		manifestClient,
+		orphanDetectionClient,
+		spec.NewResolver(keyChainLookup, extractor),
 		declarativev2.WithRemoteTargetCluster(
 			func(_ context.Context, _ declarativev2.Object) (*declarativev2.ClusterInfo, error) {
 				return &declarativev2.ClusterInfo{Config: authUser.Config()}, nil
 			},
-		), manifest.WithClientCacheKey(),
-		declarativev2.WithCustomStateCheck(statecheck.NewManagerStateCheck(statefulChecker, deploymentChecker)))
+		), manifest.WithClientCacheKey(), declarativev2.WithCustomStateCheck(declarativev2.NewExistsStateCheck()),
+	)
 
 	err = ctrl.NewControllerManagedBy(mgr).
 		For(&v1beta2.Manifest{}).
@@ -185,6 +184,7 @@ var _ = BeforeSuite(func() {
 			},
 		).Complete(reconciler)
 	Expect(err).ToNot(HaveOccurred())
+
 	Eventually(CreateNamespace, Timeout, Interval).
 		WithContext(ctx).
 		WithArguments(kcpClient, ControlPlaneNamespace).Should(Succeed())
